increased the size of the minimal accepted face

+++ tests/wheels_movements_continuous.cpp
@@ -30,6 +30,9 @@ typedef struct _CENTER_POINT
 #define MOTOR_HEAD_HORIZONTAL 0
 #define MOTOR_HEAD_VERTICAL 1
 
+#define MOTOR_FOOT_LEFT 0
+#define MOTOR_FOOT_RIGHT 1
+
 #define NUM_SECONDS_TO_WAIT_FOR_CONNECTION 3
 
 #define TOLERANCE 20
@@ -57,7 +60,7 @@ bool biggest_face(std::vector<Rect> faces, CENTER_POINT &center)
 	return found_one;
 }
 //----------------------------------------------------------------
-bool init(t_jenny5_command_module &head_controller, VideoCapture &head_cam, CascadeClassifier &face_classifier, char* error_string)
+bool init(t_jenny5_command_module &head_controller, t_jenny5_command_module &foot_controller, VideoCapture &head_cam, CascadeClassifier &face_classifier, char* error_string)
 {
 	//-------------- START INITIALIZATION ------------------------------
 
@@ -65,24 +68,43 @@ bool init(t_jenny5_command_module &head_controller, VideoCapture &head_cam, Casc
 		sprintf(error_string, "Error attaching to Jenny 5' head!");
 		return false;
 	}
+
+	if (!foot_controller.connect(11, 115200)) {
+		sprintf(error_string, "Error attaching to Jenny 5' foot!");
+		return false;
+	}
+
 	// now wait to see if I have been connected
 	// wait for no more than 3 seconds. If it takes more it means that something is not right, so we have to abandon it
 	clock_t start_time = clock();
+	bool head_responded = false;
+	bool foot_responded = false;
 
 	while (1) {
-		if (!head_controller.update_commands_from_serial())
+		if (!head_controller.update_commands_from_serial() && !foot_controller.update_commands_from_serial())
 			Sleep(5); // no new data from serial ... we make a little pause so that we don't kill the processor
 
-		if (head_controller.query_for_event(IS_ALIVE_EVENT, 0)) { // have we received the event from Serial ?
+		if (!head_responded)
+			if (head_controller.query_for_event(IS_ALIVE_EVENT, 0))  // have we received the event from Serial ?
+				head_responded = true;
+		
+		if (!foot_responded)
+			if (foot_controller.query_for_event(IS_ALIVE_EVENT, 0))  // have we received the event from Serial ?
+				foot_responded = true;
+
+		if (head_responded && foot_responded)
 			break;
-		}
 		// measure the passed time 
 		clock_t end_time = clock();
 
 		double wait_time = (double)(end_time - start_time) / CLOCKS_PER_SEC;
 		// if more than 3 seconds then game over
 		if (wait_time > NUM_SECONDS_TO_WAIT_FOR_CONNECTION) {
-			sprintf(error_string, "Head does not respond! Game over!");
+			if (!head_responded)
+			  sprintf(error_string, "Head does not respond! Game over!");
+
+			if (!foot_responded)
+				sprintf(error_string, "Foot does not respond! Game over!");
 			return false;
 		}
 	}
@@ -104,7 +126,7 @@ bool init(t_jenny5_command_module &head_controller, VideoCapture &head_cam, Casc
 	return true;
 }
 //----------------------------------------------------------------
-bool setup(t_jenny5_command_module &head_controller, char* error_string)
+bool setup(t_jenny5_command_module &head_controller, t_jenny5_command_module &foot_controller, char* error_string)
 {
 	
 	int head_motors_step_pins[2] = { 3, 6 };
@@ -112,42 +134,63 @@ bool setup(t_jenny5_command_module &head_controller, char* error_string)
 	int head_motors_enable_pins[2] = { 4, 7 };
 	head_controller.send_create_motors(2, head_motors_dir_pins, head_motors_step_pins, head_motors_enable_pins);
 
+	int foot_motors_dir_pins[4] = { 2, 5, 8, 11};
+	int foot_motors_step_pins[4] = { 3, 6, 9, 12};
+	int foot_motors_enable_pins[4] = { 4, 7, 10, 13};
+	foot_controller.send_create_motors(4, foot_motors_dir_pins, foot_motors_step_pins, foot_motors_enable_pins);
+
 	clock_t start_time = clock();
+	bool head_responded = false;
+	bool foot_responded = false;
 
 	while (1) {
-		if (!head_controller.update_commands_from_serial())
+		if (!head_controller.update_commands_from_serial() && !foot_controller.update_commands_from_serial())
 			Sleep(5); // no new data from serial ... we make a little pause so that we don't kill the processor
 
-		if (head_controller.query_for_event(MOTOR_CONTROLLER_CREATED_EVENT, 0)) { // have we received the event from Serial ?
+		if (!head_responded)
+			if (head_controller.query_for_event(MOTOR_CONTROLLER_CREATED_EVENT, 0))  // have we received the event from Serial ?
+				head_responded = true;
+
+		if (!foot_responded)
+			if (foot_controller.query_for_event(MOTOR_CONTROLLER_CREATED_EVENT, 0))  // have we received the event from Serial ?
+				foot_responded = true;
+
+		if (head_responded && foot_responded)
 			break;
-		}
+
 		// measure the passed time 
 		clock_t end_time = clock();
 
 		double wait_time = (double)(end_time - start_time) / CLOCKS_PER_SEC;
 		// if more than 3 seconds then game over
 		if (wait_time > NUM_SECONDS_TO_WAIT_FOR_CONNECTION) {
-			sprintf(error_string, "Cannot create head's motor controller! Game over!");
+			if (!head_responded)
+				sprintf(error_string, "Cannot create head's motor controller! Game over!");
+
+			if (!foot_responded)
+				sprintf(error_string, "Cannot create foot's motor controller! Game over!");
+
 			return false;
 		}
 	}
 	
-
 	head_controller.send_set_motor_speed_and_acceleration(MOTOR_HEAD_HORIZONTAL, 1000, 50);
 	head_controller.send_set_motor_speed_and_acceleration(MOTOR_HEAD_VERTICAL, 1000, 50);
 
+	foot_controller.send_set_motor_speed_and_acceleration(MOTOR_FOOT_LEFT, 500, 500);
+	foot_controller.send_set_motor_speed_and_acceleration(MOTOR_FOOT_RIGHT, 500, 500);
 	return true;
 }
 //----------------------------------------------------------------
-int	main(int argc, const char** argv)
+int	main(void)
 {
-	t_jenny5_command_module head_controller;
+	t_jenny5_command_module head_controller, foot_controller;
 	VideoCapture head_cam;
 	CascadeClassifier face_classifier;
 
 	// initialization
 	char error_string[1000];
-	if (!init(head_controller, head_cam, face_classifier, error_string)) {
+	if (!init(head_controller, foot_controller, head_cam, face_classifier, error_string)) {
 		printf("%s\n", error_string);
 		printf("Press Enter...");
 		getchar();
@@ -157,7 +200,7 @@ int	main(int argc, const char** argv)
 		printf("Initialization succceded.\n");
 
 	// setup
-	if (!setup(head_controller, error_string)) {
+	if (!setup(head_controller, foot_controller, error_string)) {
 		printf("%s\n", error_string);
 		printf("Press Enter...");
 		getchar();
@@ -185,7 +228,7 @@ int	main(int argc, const char** argv)
 		std::vector<Rect> faces;// create an array to store the faces found
 
 		// find and store the faces
-		face_classifier.detectMultiScale(grayFrame, faces, 1.1, 3, CV_HAAR_FIND_BIGGEST_OBJECT | CV_HAAR_SCALE_IMAGE, Size(50, 50));
+		face_classifier.detectMultiScale(grayFrame, faces, 1.1, 3, CV_HAAR_FIND_BIGGEST_OBJECT | CV_HAAR_SCALE_IMAGE, Size(30, 30));
 
 		CENTER_POINT center;
 
@@ -211,20 +254,20 @@ int	main(int argc, const char** argv)
 			// send a command to the module so that the face is in the center of the image
 			if (center.x < frame.cols / 2 - TOLERANCE) {
 				tracking_data angle_offset = get_offset_angles(920, Point(center.x, center.y));
-				int num_steps_x = angle_offset.grades_from_center_x / 1.8 * 16.0;
+				int num_steps_x = angle_offset.grades_from_center_x / 1.8 * 8;
 
-				head_controller.send_move_motor(MOTOR_HEAD_HORIZONTAL, -num_steps_x);
-				head_controller.set_motor_state(MOTOR_HEAD_HORIZONTAL, COMMAND_SENT);
-				printf("M%d %d# - sent\n", MOTOR_HEAD_HORIZONTAL, num_steps_x);
+				foot_controller.send_move_motor(MOTOR_FOOT_LEFT, -num_steps_x);
+				foot_controller.set_motor_state(MOTOR_FOOT_LEFT, COMMAND_SENT);
+				printf("foot: M%d %d# - sent\n", MOTOR_FOOT_LEFT, num_steps_x);
 			}
 			else
 				if (center.x > frame.cols / 2 + TOLERANCE) {
 					tracking_data angle_offset = get_offset_angles(920, Point(center.x, center.y));
-					int num_steps_x = angle_offset.grades_from_center_x / 1.8 * 16.0;
+					int num_steps_x = angle_offset.grades_from_center_x / 1.8 * 8;
 
-					head_controller.send_move_motor(MOTOR_HEAD_HORIZONTAL, -num_steps_x);
-					head_controller.set_motor_state(MOTOR_HEAD_HORIZONTAL, COMMAND_SENT);
-					printf("M%d %d# - sent\n", MOTOR_HEAD_HORIZONTAL, num_steps_x);
+					foot_controller.send_move_motor(MOTOR_FOOT_RIGHT, -num_steps_x);
+					foot_controller.set_motor_state(MOTOR_FOOT_RIGHT, COMMAND_SENT);
+					printf("foot: M%d %d# - sent\n", MOTOR_FOOT_RIGHT, num_steps_x);
 				}
 				else {
 					// face is in the center, so I do not move
@@ -240,7 +283,7 @@ int	main(int argc, const char** argv)
 
 				head_controller.send_move_motor(MOTOR_HEAD_VERTICAL, -num_steps_y);
 				head_controller.set_motor_state(MOTOR_HEAD_VERTICAL, COMMAND_SENT);
-				printf("M%d %d# - sent\n", MOTOR_HEAD_VERTICAL, num_steps_y);
+				printf("head: M%d %d# - sent\n", MOTOR_HEAD_VERTICAL, num_steps_y);
 			}
 			else
 				if (center.y > frame.rows / 2 + TOLERANCE) {
@@ -249,34 +292,42 @@ int	main(int argc, const char** argv)
 
 					head_controller.send_move_motor(MOTOR_HEAD_VERTICAL, -num_steps_y);
 					head_controller.set_motor_state(MOTOR_HEAD_VERTICAL, COMMAND_SENT);
-					printf("M%d -%d# - sent\n", MOTOR_HEAD_VERTICAL, num_steps_y);
+					printf("head: M%d -%d# - sent\n", MOTOR_HEAD_VERTICAL, num_steps_y);
 				}
 
 		}
 
-		// now extract the executed moves from the queue ... otherwise they will just stay there
+		// now extract the executed moves from the queue ... otherwise they will just sit there and will occupy memory
 		if (head_controller.get_motor_state(MOTOR_HEAD_HORIZONTAL) == COMMAND_SENT) {// if a command has been sent
-
 			if (head_controller.query_for_event(MOTOR_DONE_EVENT, MOTOR_HEAD_HORIZONTAL)) { // have we received the event from Serial ?
 				head_controller.set_motor_state(MOTOR_HEAD_HORIZONTAL, COMMAND_DONE);
 				printf("M%d# - done\n", MOTOR_HEAD_HORIZONTAL);
-			}
-			else {
-				// motor is still running and we can supervise it (with a camera?)
 			}
 		}
 
 		// now extract the moves done from the queue
 		if (head_controller.get_motor_state(MOTOR_HEAD_VERTICAL) == COMMAND_SENT) {// if a command has been sent
-
 			if (head_controller.query_for_event(MOTOR_DONE_EVENT, MOTOR_HEAD_VERTICAL)) { // have we received the event from Serial ?
 				head_controller.set_motor_state(MOTOR_HEAD_VERTICAL, COMMAND_DONE);
 				printf("M%d# - done\n", MOTOR_HEAD_VERTICAL);
 			}
-			else {
-				// motor is still running and we can supervise it (with a camera?)
+		}
+		//extract movements for foot
+		// now extract the moves done from the queue
+		if (foot_controller.get_motor_state(MOTOR_FOOT_LEFT) == COMMAND_SENT) {// if a command has been sent
+			if (foot_controller.query_for_event(MOTOR_DONE_EVENT, MOTOR_FOOT_LEFT)) { // have we received the event from Serial ?
+				foot_controller.set_motor_state(MOTOR_FOOT_LEFT, COMMAND_DONE);
+				printf("foot: M%d# - done\n", MOTOR_FOOT_LEFT);
 			}
 		}
+		if (foot_controller.get_motor_state(MOTOR_FOOT_RIGHT) == COMMAND_SENT) {// if a command has been sent
+			if (foot_controller.query_for_event(MOTOR_DONE_EVENT, MOTOR_FOOT_RIGHT)) { // have we received the event from Serial ?
+				foot_controller.set_motor_state(MOTOR_FOOT_RIGHT, COMMAND_DONE);
+				printf("foot: M%d# - done\n", MOTOR_FOOT_RIGHT);
+			}
+		}
+
+
 		if (waitKey(1) >= 0)  // break the loop
 			active = false;
 	}
